reward/requirements.txt
asgiref==3.7.2
Django==4.2.8
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==7.0.0
kafka-python==2.0.2
Markdown==3.5.1
psycopg==3.1.15
psycopg-binary==3.1.15
psycopg-pool==3.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
zipp==3.17.0
